Fix space leak in encodeMessages
There seems to be a space leak related to the use of atomicModifyIORef (i.e. the non-strict variant) in the context of encoding messages in Network.WebSockets.Hybi13. Specifically, in a server application, it appears that gen' is never forced, thus building up a chain of thunks during the lifetime of a connection as messages are sent by the server. The following results are from a test setup with 1000 clients sending a ping every second (to which the server responds with a pong) for a duration of 5 minutes. I also attached the complete sample project, just in case.