Fix runClientWithSocket exception behaviour
In some cases runClientWithSocket will throw a different exception than
the one raised in its `app` argument.

Because of 'bracket' it will always attempt to close the stream when any
exception is thrown by `runClientWithStream`.

Consider the scenario where a thread is waiting on data. Ultimately it
blocks on receive' from makeStream. If you send an asynchronous
exception (like ThreadKilled) to this thread the stream will be closed
because of `onException` in receive'.

Next runClientWithSocket will call `Stream.close`. This attempts to send
some message over the websocket. This in turn throws the
`ConnectionClosed` exception. But we where already in an exception
handler to begin with! So now the original exception is masked by
`ConnectionClosed`.